Split shared/types into domain modules behind a barrel
Break the ~988-line src/shared/types.ts (74 exports) into ten
domain-scoped modules under src/shared/types/ (common, project, layout,
settings, git, terminal, logs, task-runs, search, tree), re-exported from
a thin barrel at src/shared/types/index.ts so every existing @shared/types
and ../shared/types import resolves unchanged. Peel the renderer view-state
group (WorkingSurface/LeftView/LayoutState) into layout.ts, kept under
shared/ because the layout is persisted through the settings IPC contract
(getLayout/setLayout, Settings.layout, the native-menu checkbox sync).

Mechanical, non-breaking: no type shapes or values changed. Update AGENTS.md,
docs/, and three source comments that referenced the old file path.
